style: replace all u64 with usize; from now on, we will never use u64.

+++ os/src/syscall/mod.rs
@@ -3,9 +3,9 @@ mod syscall;
 pub use syscall::*;
 use crate::driver::shutdown;
 
-pub unsafe fn syscall(index : u64 , a0 : u64, a1 : u64, a2 : u64) {
+pub unsafe fn syscall(index : usize , a0 : usize, a1 : usize, a2 : usize) {
     use sys::syscall::*;
-    let _ = index as u64 + a0 + a1 + a2;
+    let _ = index as usize + a0 + a1 + a2;
     match index {
         SYS_SHUTDOWN => shutdown(),
         SYS_SLEEP    => sys_yield(),
