+++ os/src/trap/user.rs
@@ -78,12 +78,12 @@ pub unsafe fn user_trap_return() {
 
 unsafe fn return_to_user(base : PageAddress) {
     let satp = base.bits() | (8 << 60); // Sv39
-    let func = TRAMPOLINE + (user_return as u64 - user_handle as u64);
+    let func = TRAMPOLINE + (user_return as usize - user_handle as usize);
 
     message!("Returning to user space with satp: {:#x}", satp);
     message!("Returning to user space with addr: {:#x}", func);
 
-    type CallType = fn(u64);
+    type CallType = fn(usize);
     let ptr  = &func as *const _;
     let ptr  = ptr as *const CallType;
